Remove the savePermission on model creation.
This is not needed anymore since Daybed supports it directly. For more
information, see the bug on the daybed side:
https://github.com/spiral-project/daybed/pull/230

I removed bot the test and the if block since they're not doing much. I don't
believe we need to test anything specific here since the logic is in daybed,
not daybed.js.